+++ src/components/Auth/SignInWithOAuth/SignInWithOAuth.jsx
@@ -1,5 +1,5 @@
-import { OAUTH_PROVIDERS } from "../../constants/constants";
-import useSignInWithOAuth from "../../hooks/useSignInWithOAuth";
+import { OAUTH_PROVIDERS } from "../../../config/constants/constants";
+import useSignInWithOAuth from "../../../hooks/useSignInWithOAuth";
 
 const SignInWithOAuth = () => {
   const { handleSignInWithOAuth } = useSignInWithOAuth();
